nmad:   -- remove the global event monitor in sendrecv
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@26438 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ nmad/interfaces/sendrecv/include/nm_sendrecv_interface.h
@@ -31,7 +31,7 @@
 
 
 /** a request status or event */
-typedef uint8_t nm_sr_status_t;
+typedef uint16_t nm_sr_status_t;
 
 /** a posted send has completed */
 #define NM_SR_STATUS_SEND_COMPLETED  ((nm_sr_status_t)NM_STATUS_PACK_COMPLETED)
